+++ Math/Problem7/main.cpp
@@ -5,12 +5,12 @@
 
 using namespace std;
 
-unsigned long long divisorSum( unsigned long long input )
+unsigned long divisorSum( unsigned long input )
 {
-    unsigned long long sum = 1;
-    unsigned long long max = sqrt(input) ;
+    unsigned long sum = 1;
+    unsigned long max = sqrt(input) ;
 
-    for( unsigned long long i = 2 ; i <= max ; i++ )
+    for( unsigned long i = 2 ; i <= max ; i++ )
     {
         if( input % i == 0 )
         {
@@ -29,8 +29,8 @@ unsigned long long divisorSum( unsigned long long input )
 
 int main(int argc, char* argv[])
 {
-    unsigned long long i,j ;
-    unsigned long long skip[SKIP_ARR_SIZE]={0,0,0,0,0,0,0};
+    unsigned long i,j ;
+    unsigned long skip[SKIP_ARR_SIZE]={0,0,0,0,0,0,0};
 
     for( i = 1 ; i <= END_NUM ; i++ )
     {
